+++ frontend/src/components/ui/Loader.tsx
@@ -1,16 +1,21 @@
 import React from "react";
 interface LoaderProps {
   size?: `small` | `medium` | `large` | `xl`;
+  variant?: `standard` | `button`;
 }
 
-const Loader: React.FC<LoaderProps> = ({ size = `medium` }) => {
+const Loader: React.FC<LoaderProps> = ({ size = `medium`, variant = `standard` }) => {
   const pxSize = size === `small` ? 12 : size === `medium` ? 16 : size === `large` ? 20 : 24;
 
   return (
     <div className="w-full h-full flex items-center justify-center">
       <div className="flex flex-col gap-5 items-center">
         <div
-          className={`rounded-full border-8 border-t-primary animate-spin w-${pxSize} h-${pxSize}`}
+          className={
+            variant === `standard`
+              ? `rounded-full  border-8 border-t-primary animate-spin w-${pxSize} h-${pxSize}`
+              : `rounded-full animate-spin border-2 border-primary-foreground border-t-transparent w-6 h-6`
+          }
         ></div>
       </div>
     </div>