+++ frontend/src/pages/UserFormPage.tsx
@@ -4,6 +4,7 @@ import { ERROR_MESSAGES } from "@/enums/ErrorMessages";
 import { MainRoutes } from "@/enums/Routes";
 import { useUsersApi } from "@/hooks/api/useUsersApi";
 import { IUser } from "@/interfaces/IUser";
+import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { useEffect, useState } from "react";
 import { useNavigate, useParams } from "react-router-dom";
 import { toast } from "sonner";
@@ -15,9 +16,11 @@ const UserFormPage = () => {
 
   const [user, setUser] = useState<IUser | null>(null);
   const [isLoading, setIsLoading] = useState(false);
+  const queryClient = useQueryClient();
 
   const onSuccess = () => {
     toast.success(`משתמש נשמר בהצלחה!`);
+    queryClient.invalidateQueries({ queryKey: [`users`] });
     navigation(MainRoutes.USERS);
   };
 
@@ -25,15 +28,23 @@ const UserFormPage = () => {
     toast.error(ERROR_MESSAGES.GENERIC_ERROR_MESSAGE, { description: e.message });
   };
 
+  const editUser = useMutation({
+    mutationFn: ({ id, user }: { id: string; user: IUser }) => updateUser(id, user),
+    onSuccess: onSuccess,
+    onError: onError,
+  });
+
+  const addNewUser = useMutation({
+    mutationFn: addUser,
+    onSuccess: onSuccess,
+    onError: onError,
+  });
+
   const handleSaveUser = (user: IUser) => {
     if (id) {
-      updateUser(id, user)
-        .then(() => onSuccess())
-        .catch(onError);
+      editUser.mutate({ id, user });
     } else {
-      addUser(user)
-        .then(() => onSuccess())
-        .catch(onError);
+      addNewUser.mutate(user);
     }
   };
 
@@ -47,12 +58,16 @@ const UserFormPage = () => {
       .finally(() => setIsLoading(false));
   }, []);
 
-  if (isLoading) return <Loader />;
+  if (isLoading) return <Loader size="large" />;
 
   return (
     <div>
       <h1 className="font-bold text-2xl">פרטי משתמש</h1>
-      <UserForm existingUser={user} saveInfo={(user) => handleSaveUser(user)} />
+      <UserForm
+        existingUser={user}
+        saveInfo={(user) => handleSaveUser(user)}
+        pending={editUser.isPending || addNewUser.isPending}
+      />
     </div>
   );
 };