+++ frontend/CHANGELOG.md
@@ -11,16 +11,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Added ability to delete a user from User's table.
 - Added ability to double click on row to view items.
+- Added custom button with loading option
+- Added button variant to Loader
 
 ### Fixed
 
-- Fixed weigh ins not being displayed properly in the calendar. 
-- Fixed multiple inproper validation checks causing issues to add/update items. 
+- Fixed weigh ins not being displayed properly in the calendar.
+- Fixed multiple inproper validation checks causing issues to add/update items.
 - Fixed issue where you can have less maxReps than minReps.
 - Fixed issue where selecting a diet plan preset would not update the page properly.
 - Fixed issue where selecting a workout plan preset would not update the page properly.
+- Fixed combobox loader not working in muscleGroup selector
 
-### Changed 
+### Changed
 
 - Diet plan now updates automatically when creating/editing.
 - Multiple UI/UX improvements.
+- implemented React Querry in most forms